theme-charcoal -- throwaway (eventually) branch
The purpose for keeping a copy of this branch is to have it available
as a reference. This branch includes some color experimentation for
the proofreading interfaces that I'd like to keep around so I can
look back at them later, but for now, things are going off in another
direction, and we will probably exclude dealing with the proofreading
interfaces and all the other page-browsing stuff until after initial
release of the charcoal theme.

Casey has merged `theme.less` into `layout.less`, and created a
separate `page_interfaces.less` file to corral all of the CSS related
to anything having to do with the proofing interfaces or project-page
navigation, including WC context checking of images, viewing images
from the Page Detail list, the mentor interface for viewing images and
text side-by-side.