Probably start this branch over again, but ...
I need to figure out what to keep. I got my wires crossed when
starting the code changes. This should only include changes to
page_interfaces.less to correct the wc stats colors and to make
tdbottom match tdtop in background color and text color. If I
do anything with charcoal.less, it should be just to make sure
that the PIs match what's on PROD, and leave any charcoal PI
theming to the other branch.